feat: refactor structure and do fine the update of plan
