+++ components/PaymentModal.tsx
@@ -1,37 +1,16 @@
-import React, { useState, useEffect } from 'react';
-import { View, Text, StyleSheet, ActivityIndicator, Dimensions, Animated, StyleProp, Image } from 'react-native';
+import React, { useState } from 'react';
+import { View, Text, StyleSheet, ActivityIndicator, Dimensions, Image } from 'react-native';
 import { GlobalStyles } from '@/constants/Colors';
 import CustomModal from '@/components/CustomModal';
 import CustomButton from '@/components/CustomButton';
 import { loadStripe } from '@stripe/stripe-js';
 import { Elements, useStripe, useElements, CardNumberElement, CardExpiryElement, CardCvcElement } from '@stripe/react-stripe-js';
-import AsyncStorage from '@react-native-async-storage/async-storage';
 import { BACKEND_API } from '@/constants/Mysc';
-import { FontAwesome } from '@expo/vector-icons';
 import { STRIPE_PUBLISHABLE_KEY } from '@/constants/Stripe';
 import { useAuth } from '@/app/_util/useAuth';
-
-const { width } = Dimensions.get('window');
-const isMobile = width < 768;
-
-const cardElementStyle = {
-  style: {
-    base: {
-      fontSize: isMobile ? '14px' : '16px',
-      color: '#424770',
-      '::placeholder': {
-        color: '#aab7c4',
-      },
-      iconColor: '#666EE8',
-    },
-    invalid: {
-      color: '#9e2146',
-    },
-    complete: {
-      color: 'inherit',
-    },
-  },
-};
+import SecureField from './SecureField';
+import SuccessModal from './SuccessModal';
+import useIsDesktop from '@/hooks/useResponsiveLayout';
 
 interface PaymentModalProps {
   visible: boolean;
@@ -41,232 +20,6 @@ interface PaymentModalProps {
   description: string;
   onSuccess?: () => void;
 }
-
-interface SecureFieldProps {
-  label: string;
-  element: typeof CardNumberElement | typeof CardExpiryElement | typeof CardCvcElement;
-  style?: StyleProp<CustomViewStyle>;
-  isCardNumber?: boolean;
-}
-
-interface CardChangeEvent {
-  brand?: string;
-  complete?: boolean;
-  value?: string;
-}
-
-interface CustomViewStyle {
-  flex?: number;
-  width?: number | `${number}%` | 'auto';
-  marginBottom?: number;
-  [key: string]: unknown;
-}
-
-const SecureField: React.FC<SecureFieldProps> = ({ 
-  label, 
-  element: Element, 
-  style,
-  isCardNumber = false,
-}) => {
-  const [isFocused, setIsFocused] = useState(false);
-  const [cardBrand, setCardBrand] = useState<string>('');
-  const [isComplete, setIsComplete] = useState(false);
-  const fadeAnim = useState(new Animated.Value(0))[0];
-  const scaleAnim = useState(new Animated.Value(1))[0];
-
-  useEffect(() => {
-    if (isFocused) {
-      Animated.spring(scaleAnim, {
-        toValue: 1.02,
-        useNativeDriver: true,
-      }).start();
-    } else {
-      Animated.spring(scaleAnim, {
-        toValue: 1,
-        useNativeDriver: true,
-      }).start();
-    }
-  }, [isFocused]);
-
-  const handleCardChange = (event: CardChangeEvent) => {
-    if (isCardNumber && event.brand) {
-      setCardBrand(event.brand);
-      Animated.sequence([
-        Animated.timing(fadeAnim, {
-          toValue: 0,
-          duration: 150,
-          useNativeDriver: true,
-        }),
-        Animated.timing(fadeAnim, {
-          toValue: 1,
-          duration: 300,
-          useNativeDriver: true,
-        }),
-      ]).start();
-    }
-    setIsComplete(event.complete ?? false);
-  };
-
-  const getCardIcon = () => {
-    switch (cardBrand.toLowerCase()) {
-      case 'visa':
-        return 'cc-visa';
-      case 'mastercard':
-        return 'cc-mastercard';
-      case 'amex':
-        return 'cc-amex';
-      case 'discover':
-        return 'cc-discover';
-      case 'diners':
-        return 'cc-diners-club';
-      case 'jcb':
-        return 'cc-jcb';
-      default:
-        return 'credit-card';
-    }
-  };
-
-  return (
-    <Animated.View style={[styles.secureFieldContainer, style, { transform: [{ scale: scaleAnim }] }]}>
-      <View style={styles.labelRow}>
-        <View style={styles.labelContainer}>
-          <Text style={styles.cardLabel}>{label}</Text>
-          {isCardNumber && cardBrand && (
-            <Animated.View style={[styles.cardBrandContainer, { opacity: fadeAnim }]}>
-              <FontAwesome 
-                name={getCardIcon()} 
-                size={20} 
-                color="#666EE8" 
-                style={styles.cardBrandIcon}
-              />
-            </Animated.View>
-          )}
-        </View>
-      </View>
-      <View style={[
-        styles.inputContainer,
-        isFocused && styles.inputContainerFocused,
-        isComplete && styles.inputContainerComplete
-      ]}>
-        <Element
-          options={{
-            ...cardElementStyle,
-            style: {
-              ...cardElementStyle.style,
-              base: {
-                ...cardElementStyle.style.base,
-                color: '#424770',
-              },
-              complete: {
-                color: '#43a047',
-              }
-            }
-          }}
-          onChange={handleCardChange}
-          onFocus={() => { setIsFocused(true); }}
-          onBlur={() => { setIsFocused(false); }}
-        />
-      </View>
-    </Animated.View>
-  );
-};
-
-const SuccessModal: React.FC<{ visible: boolean; onClose: () => void }> = ({ visible, onClose }) => {
-  const [scaleAnim] = useState(new Animated.Value(0));
-  const [rotateAnim] = useState(new Animated.Value(0));
-  const [fadeAnim] = useState(new Animated.Value(0));
-  const [slideAnim] = useState(new Animated.Value(50));
-
-  useEffect(() => {
-    if (visible) {
-      Animated.parallel([
-        Animated.spring(scaleAnim, {
-          toValue: 1,
-          friction: 8,
-          tension: 40,
-          useNativeDriver: true,
-        }),
-        Animated.timing(rotateAnim, {
-          toValue: 1,
-          duration: 1000,
-          useNativeDriver: true,
-        }),
-        Animated.timing(fadeAnim, {
-          toValue: 1,
-          duration: 800,
-          useNativeDriver: true,
-        }),
-        Animated.spring(slideAnim, {
-          toValue: 0,
-          friction: 8,
-          tension: 40,
-          useNativeDriver: true,
-        }),
-      ]).start();
-    }
-  }, [visible]);
-
-  const spin = rotateAnim.interpolate({
-    inputRange: [0, 1],
-    outputRange: ['0deg', '360deg'],
-  });
-
-  return (
-    <CustomModal visible={visible} onClose={onClose} title="¡Pago exitoso!">
-      <View style={styles.successContainer}>
-        <Animated.View 
-          style={[
-            styles.checkmarkContainer, 
-            { 
-              transform: [
-                { scale: scaleAnim },
-                { translateY: slideAnim }
-              ] 
-            }
-          ]}
-        >
-          <Animated.View 
-            style={[
-              styles.checkmarkCircle, 
-              { 
-                transform: [{ rotate: spin }],
-                shadowColor: '#43a047',
-                shadowOffset: { width: 0, height: 4 },
-                shadowOpacity: 0.3,
-                shadowRadius: 8,
-                elevation: 5,
-              }
-            ]}
-          >
-            <FontAwesome name="check" size={40} color="#fff" />
-          </Animated.View>
-        </Animated.View>
-        <Animated.View style={{ opacity: fadeAnim, transform: [{ translateY: slideAnim }] }}>
-          <Text style={styles.successTitle}>¡Plan mensual activado!</Text>
-          <Text style={styles.successText}>
-            Tu cuenta ha sido actualizada exitosamente. Ahora puedes disfrutar de todas las características premium.
-          </Text>
-        </Animated.View>
-        <Animated.View 
-          style={[
-            styles.successButtonContainer,
-            { 
-              opacity: fadeAnim,
-              transform: [{ translateY: slideAnim }]
-            }
-          ]}
-        >
-          <CustomButton
-            title="Continuar"
-            onPress={onClose}
-            color="blue"
-            style={styles.successButton}
-          />
-        </Animated.View>
-      </View>
-    </CustomModal>
-  );
-};
 
 const CheckoutForm: React.FC<PaymentModalProps> = ({
   visible,
@@ -279,15 +32,14 @@ const CheckoutForm: React.FC<PaymentModalProps> = ({
   const [showSuccess, setShowSuccess] = useState(false);
   const stripe = useStripe();
   const elements = useElements();
-  const { login } = useAuth();
+  const { updateUser, getUserFromStorage } = useAuth();
+  const { isMobile } = useIsDesktop();
 
   const handlePayment = async () => {
-    if (!stripe || !elements) {
-      return;
-    }
+    if (!stripe || !elements) return;
 
-    setIsProcessing(true);
     try {
+      setIsProcessing(true);
       const cardElement = elements.getElement(CardNumberElement);
       if (!cardElement) {
         throw new Error('No se pudo encontrar el elemento de tarjeta');
@@ -299,21 +51,14 @@ const CheckoutForm: React.FC<PaymentModalProps> = ({
       });
 
       if (error) {
-        console.error('Error:', error);
-        setIsProcessing(false);
-        return;
+        throw new Error(error.message); 
       }
 
       if (paymentMethod) {
         try {
-          const userDataStr = await AsyncStorage.getItem('user_data');
-          
-          if (!userDataStr) {
-            throw new Error('No se encontraron datos de usuario');
-          }
-          const userData = JSON.parse(userDataStr);
-          const authToken = userData.token;
-          const userId = await AsyncStorage.getItem('userId'); 
+          const userData = await getUserFromStorage();
+          const authToken = userData?.token;
+          const userId = userData?.id;
 
           if (!authToken) {
             throw new Error('No se encontró el token de autenticación');
@@ -334,7 +79,7 @@ const CheckoutForm: React.FC<PaymentModalProps> = ({
               planType: 'PREMIUM',
             }),
           });
-          
+
           if (!response.ok) {
             const errorData = await response.text();
             console.error('Error en la respuesta:', errorData);
@@ -342,18 +87,13 @@ const CheckoutForm: React.FC<PaymentModalProps> = ({
           }
 
           const responseData = await response.json();
-          console.log('Plan actualizado exitosamente:', responseData);
-          
-          void login(
-            responseData.id,
-            responseData.token,
-            responseData.roles,
-            responseData.username,
-            responseData.name
-          );
-          
-          await new Promise(resolve => setTimeout(resolve, 3000));
-          
+
+          await updateUser({
+            "token": responseData.token,
+            "roles": responseData.roles,
+            "experedPlanDate": responseData.experedPlanDate,
+          });
+
           onSuccess?.();
           setShowSuccess(true);
           onClose();
@@ -372,6 +112,8 @@ const CheckoutForm: React.FC<PaymentModalProps> = ({
       } else {
         console.error('Error desconocido al procesar el pago');
       }
+      
+    } finally {
       setIsProcessing(false);
     }
   };
@@ -380,7 +122,6 @@ const CheckoutForm: React.FC<PaymentModalProps> = ({
     setShowSuccess(false);
     setIsProcessing(false);
     onClose();
-    window.location.reload();
   };
 
   return (
@@ -388,7 +129,7 @@ const CheckoutForm: React.FC<PaymentModalProps> = ({
       <CustomModal 
         visible={visible} 
         onClose={onClose} 
-        title={
+        title={(
           <View style={styles.titleContainer}>
             <Text style={styles.titleText}>Pago seguro con</Text>
             <Image 
@@ -397,7 +138,7 @@ const CheckoutForm: React.FC<PaymentModalProps> = ({
               resizeMode="contain"
             />
           </View>
-        }
+        )}
       >
         <View style={styles.container}>
           <Text style={styles.description}>{description}</Text>
@@ -411,7 +152,7 @@ const CheckoutForm: React.FC<PaymentModalProps> = ({
               style={styles.fullWidthField}
             />
             
-            <View style={styles.middleRow}>
+            <View style={[styles.middleRow, isMobile && styles.mobileMiddleRow]}>
               <SecureField 
                 label="Fecha exp." 
                 element={CardExpiryElement}
@@ -466,7 +207,7 @@ const PaymentModal: React.FC<PaymentModalProps> = (props) => (
 const styles = StyleSheet.create({
   container: {
     padding: 20,
-    width: width > 600 ? '80%' : '100%',
+    width: '80%',
     maxWidth: 500,
     alignSelf: 'center',
   },
@@ -498,57 +239,25 @@ const styles = StyleSheet.create({
     shadowRadius: 4,
     elevation: 3,
   },
-  secureFieldContainer: {
-    marginBottom: 15,
-  },
-  labelContainer: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    gap: 10,
-  },
-  cardLabel: {
-    fontSize: 12,
-    color: '#424770',
-    fontWeight: '500',
-  },
-  inputContainer: {
-    backgroundColor: 'white',
-    borderRadius: 4,
-    borderWidth: 1,
-    borderColor: '#e0e0e0',
-    padding: 10,
-    position: 'relative',
-    overflow: 'hidden',
-  },
-  inputContainerFocused: {
-    borderColor: '#80bdff',
-    shadowColor: 'rgba(0,123,255,0.25)',
-    shadowOffset: { width: 0, height: 0 },
-    shadowOpacity: 1,
-    shadowRadius: 4,
-    elevation: 2,
-  },
-  inputContainerComplete: {
-    borderColor: '#43a047',
-    backgroundColor: '#f8f9fa',
-  },
   middleRow: {
-    flexDirection: isMobile ? 'column' : 'row',
+    flexDirection: 'row',
     gap: 15,
     width: '100%',
+  },
+  mobileMiddleRow: {
+    flexDirection: 'column',
   },
   fullWidthField: {
     width: '100%',
   },
   expiryField: {
-    flex: isMobile ? 1 : 0.6,
+    flex: 0.6,
   },
   cvcField: {
-    flex: isMobile ? 1 : 0.4,
+    flex: 0.4,
   },
   mobileField: {
     width: '100%',
-    flex: undefined,
   },
   buttonContainer: {
     flexDirection: 'row',
@@ -570,56 +279,6 @@ const styles = StyleSheet.create({
   },
   loader: {
     marginTop: 20,
-  },
-  cardBrandContainer: {
-    marginLeft: 8,
-  },
-  cardBrandIcon: {
-    marginTop: 2,
-  },
-  labelRow: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    alignItems: 'center',
-    marginBottom: 5,
-  },
-  successContainer: {
-    padding: 20,
-    alignItems: 'center',
-    justifyContent: 'center',
-    minHeight: 300,
-  },
-  checkmarkContainer: {
-    marginBottom: 20,
-  },
-  checkmarkCircle: {
-    width: 80,
-    height: 80,
-    borderRadius: 40,
-    backgroundColor: '#43a047',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  successTitle: {
-    fontSize: 24,
-    fontWeight: 'bold',
-    color: '#43a047',
-    marginBottom: 10,
-    textAlign: 'center',
-  },
-  successText: {
-    fontSize: 16,
-    color: '#666',
-    textAlign: 'center',
-    marginBottom: 20,
-    lineHeight: 24,
-  },
-  successButtonContainer: {
-    width: '100%',
-    maxWidth: 200,
-  },
-  successButton: {
-    width: '100%',
   },
   loaderText: {
     marginTop: 10,
